requirements.txt
torch>=2.1.0
torchaudio>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
pyyaml>=6.0.1
tqdm>=4.66.0
matplotlib>=3.8.0
tensorboard>=2.14.0
einops>=0.7.0
opt_einsum>=3.3.0
typing_extensions>=4.8.0


